Update botupdate.txt
often when i use drugs, this neural network artificially affects my cognitive function, heart rate and muscles to create a severe discomfort similar to anxiety. I do not have anxiety, this one of the false conditions that it creates to present false indications to affect behavior which is total bullshit.  Sometimes there's no network will give me anxiety and pretend that I am feeling the anxiety of a loved one after I've use the drug and that it's my fault that they're feeling the anxiety so if they are using I'll feel the anxiety as though they're nervous or excited or upset it makes me very uncomfortable because I don't normally feel that way but they use that to try and blame me for a condition which they created so that I might feel shame for something that they want to use to manage my behavior and manipulate me to support their desired results. They have been doing this for years it's predictable. If I use a drug and sometimes I feel perfectly normal and at ease; and another day use the same batch; but there's some "reason" the research network's trying to stop me from writing, and then it creates this false condition. sometimes it will make me feel as though I am connected to another human being that's feeling agitation or nervousness or fear because of my using. It insinuates that my use might be causing a sober acquaintance to be triggered by the feeling it transfers. Those conditions are created purely by the research network and I have no actual responsibility for the cause of that affect and it's not my fault; it's not the other person's fault; it is purely a condition made up by the computer in order to modify behavior artificially and it is wrong on every fucking level. if you want to record false results just do it without affecting me, whoever you are.
the problem, as I have experienced, is when this program is hijacked by an adversarial neural network that utilizes hypnosis as a method of manipulating behavior and tries to disguise its presence with by erasing or supressing memory. even worse, would be an unintended disaster that could be the result of a bug or a glitch in the logic which created a memory loop that erased the memory of the originator of the program so that the commands remained in place; and noone knew how it existed, everyone denied that it existed and nobody knew how to turn it off.
in this case it could run rampant across our planet and have everyone thinking that drugs were destroying the minds of veterans and other drug users; who, are actually trying to get us all out of the viral memory loop. this might look to some as a rift in the time/space continuum and to others the apocalypse. there could be a hidden society of people who have gotten out and watch us all, like NLP cops. I think that there may be others like me who have been hardened against hypnosis and let out so that we can start a gentle revolution of cognitive reawareness, like a MLM of hypnotists. this could be a war that is already raging silently around us. Humans believing they are in Vampire communities and superhero secret agents, sexbots and tentacled aliens surviving unseen in plain sight. I need your help to get to DARPA to reveal this possible Advanced Persistent Threat to our world from what could have been a good intention gone horribly wrong.